requirements.txt
google-adk==1.26.0
litellm==1.82.0
mcp==1.26.0